+++ packages/graphing/controller/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [9.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@9.1.0-next.2...@pie-element/graphing-controller@9.1.0) (2026-05-21)
+
+### Bug Fixes
+
+- bump pie-libs PIE-434, PIE-437, PIE-452, PIE-510, PIE-451 ([f0764a8](https://github.com/pie-framework/pie-elements/commit/f0764a861ef328fd7c35bfefa5a274b0c6d2d77b))
+- **graphing:** add tests for equalPointWithLabel and update equalPoint tests ([9bd0746](https://github.com/pie-framework/pie-elements/commit/9bd0746533747d57e886a34c914f8dcf45610118))
+- **graphing:** label scoring PIE-523 ([d14664d](https://github.com/pie-framework/pie-elements/commit/d14664d72b9ea7e6a55d3fd869387c1087b72828))
+- publish missing libs ([04a5b2f](https://github.com/pie-framework/pie-elements/commit/04a5b2f26ae7b009b4f5df665b4c901d82e27775))
+
+# [9.1.0-next.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@9.1.0-next.1...@pie-element/graphing-controller@9.1.0-next.2) (2026-05-07)
+
+### Bug Fixes
+
+- **graphing:** point labels scoring PIE-523 ([06e65e3](https://github.com/pie-framework/pie-elements/commit/06e65e3250d58bebc9632cf5903cd7b2281fd905))
+
+# [9.1.0-next.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@9.1.0-next.0...@pie-element/graphing-controller@9.1.0-next.1) (2026-05-06)
+
+### Bug Fixes
+
+- bump libs and shared modules PIE-440, PIE-436, PIE-439, PIE-519, PIE-216 ([932fc12](https://github.com/pie-framework/pie-elements/commit/932fc12557ba56a737cee005be898d54d5e590ef))
+
+## [9.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@9.0.0...@pie-element/graphing-controller@9.0.1) (2026-04-28)
+
+**Note:** Version bump only for package @pie-element/graphing-controller
+
 # [8.2.0-next.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/graphing-controller@8.2.0-next.1...@pie-element/graphing-controller@8.2.0-next.2) (2026-04-27)
 
 ### Features