+++ packages/multiple-choice/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [13.2.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@13.2.0-next.1...@pie-element/multiple-choice@13.2.0) (2026-05-21)
+
+### Bug Fixes
+
+- bump pie-libs PIE-434, PIE-437, PIE-452, PIE-510, PIE-451 ([f0764a8](https://github.com/pie-framework/pie-elements/commit/f0764a861ef328fd7c35bfefa5a274b0c6d2d77b))
+- bump shared modules and libs PIE-522, PIE-451, PIE-511, PIE-561, PIE-562, PIE-563 ([5cfc2e9](https://github.com/pie-framework/pie-elements/commit/5cfc2e9f7d6c1a485538fd674f8602dd92018c50))
+- **multiple-choice:** decouple radio group name from partLabel to keep EBSR parts in separate radio groups regardless of partLabels setting PIE-174 ([73b622d](https://github.com/pie-framework/pie-elements/commit/73b622d996f1e2e21091f3cf80f819856049fe90))
+- **multiple-choice:** fix choice mode button PIE-560 ([98cce22](https://github.com/pie-framework/pie-elements/commit/98cce22ad5a3a295571a4bc493f0eb476ef2fb71))
+- **multiple-choice:** forward autoFocusRef via inputRef so Tab lands on the first radio in a group, use a unique random radio group name per instance so EBSR parts stay independent regardless of partLabels PIE-174 ([9ed65f5](https://github.com/pie-framework/pie-elements/commit/9ed65f5904e23f035f07594ddcb5ea5a6ff0a5aa))
+- publish missing libs ([04a5b2f](https://github.com/pie-framework/pie-elements/commit/04a5b2f26ae7b009b4f5df665b4c901d82e27775))
+- test ([e0a769d](https://github.com/pie-framework/pie-elements/commit/e0a769d8c4411b2afa08b0ebd48da24154302f8a))
+
+# [13.2.0-next.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@13.2.0-next.0...@pie-element/multiple-choice@13.2.0-next.1) (2026-05-06)
+
+### Bug Fixes
+
+- bump libs and shared modules PIE-440, PIE-436, PIE-439, PIE-519, PIE-216 ([932fc12](https://github.com/pie-framework/pie-elements/commit/932fc12557ba56a737cee005be898d54d5e590ef))
+- fixed tests [PIE-439] ([aef44f6](https://github.com/pie-framework/pie-elements/commit/aef44f686c0d69c83f3fe464509092d68b32cbc1))
+- make sure initial changes to the model are applied to what is needed [PIE-439] ([892a3cc](https://github.com/pie-framework/pie-elements/commit/892a3cc4c6f0b37cb65f42cfdd9a9a01db447aed))
+
+# [13.1.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@13.0.1...@pie-element/multiple-choice@13.1.0) (2026-05-04)
+
+### Features
+
+- **multiple-choice / ebsr:** handle logic for player heading attributes PIE-153 PIE-154 ([3530ce6](https://github.com/pie-framework/pie-elements/commit/3530ce6e6bad9f032ceb7c774cfdc66a43c1f252))
+
+## [13.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@13.0.0...@pie-element/multiple-choice@13.0.1) (2026-04-28)
+
+**Note:** Version bump only for package @pie-element/multiple-choice
+
 # [12.2.0-next.12](https://github.com/pie-framework/pie-elements/compare/@pie-element/multiple-choice@12.2.0-next.11...@pie-element/multiple-choice@12.2.0-next.12) (2026-04-27)
 
 ### Features