+++ packages/pie-models/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.0.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@16.0.2-next.0...@pie-element/pie-models@16.0.2) (2026-05-21)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
+## [16.0.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@16.0.0...@pie-element/pie-models@16.0.1) (2026-04-28)
+
+**Note:** Version bump only for package @pie-element/pie-models
+
 ## [15.1.1-next.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/pie-models@15.1.1-next.1...@pie-element/pie-models@15.1.1-next.2) (2026-04-03)
 
 **Note:** Version bump only for package @pie-element/pie-models